+++ av1an-core/src/chunk.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use std::ffi::OsString;
 use std::path::Path;
 
@@ -88,183 +91,6 @@ impl Chunk {
       insert_noise_table_params(self.encoder, &mut self.video_params, &grain_table)?;
     }
 
-    Ok(())
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-  use std::path::PathBuf;
-
-  #[test]
-  fn test_chunk_name_1() {
-    let ch = Chunk {
-      temp: "none".to_owned(),
-      index: 1,
-      input: Input::Video {
-        path: "test.mkv".into(),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 0,
-      end_frame: 5,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::x264,
-      noise_size: (None, None),
-      ignore_frame_mismatch: false,
-    };
-    assert_eq!("00001", ch.name());
-  }
-  #[test]
-  fn test_chunk_name_10000() {
-    let ch = Chunk {
-      temp: "none".to_owned(),
-      index: 10000,
-      input: Input::Video {
-        path: "test.mkv".into(),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 0,
-      end_frame: 5,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::x264,
-      noise_size: (None, None),
-      ignore_frame_mismatch: false,
-    };
-    assert_eq!("10000", ch.name());
-  }
-
-  #[test]
-  fn test_chunk_output() {
-    let ch = Chunk {
-      temp: "d".to_owned(),
-      index: 1,
-      input: Input::Video {
-        path: "test.mkv".into(),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 0,
-      end_frame: 5,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::x264,
-      noise_size: (None, None),
-      ignore_frame_mismatch: false,
-    };
-    assert_eq!("d/encode/00001.ivf", ch.output());
-  }
-
-  #[test]
-  fn test_chunk_frames() {
-    let ch = Chunk {
-      temp: "none".to_owned(),
-      index: 1,
-      input: Input::Video {
-        path: "test.mkv".into(),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 10,
-      end_frame: 25,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::x264,
-      noise_size: (None, None),
-      ignore_frame_mismatch: false,
-    };
-    assert_eq!(15, ch.frames());
-  }
-
-  #[test]
-  fn test_apply_photon_noise_args_with_noise() -> anyhow::Result<()> {
-    let temp_dir = tempfile::tempdir()?;
-    let mut ch = Chunk {
-      temp: temp_dir.path().to_str().unwrap().to_owned(),
-      index: 1,
-      input: Input::Video {
-        path: PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-files/blank_1080p.mkv"),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 0,
-      end_frame: 5,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::svt_av1,
-      noise_size: (Some(1920), Some(1080)),
-      ignore_frame_mismatch: false,
-    };
-
-    ch.apply_photon_noise_args(Some(8), true)?;
-    assert!(ch.video_params.iter().any(|p| p.contains("fgs-table")));
-    Ok(())
-  }
-
-  #[test]
-  fn test_apply_photon_noise_args_no_noise() -> anyhow::Result<()> {
-    let temp_dir = tempfile::tempdir()?;
-    let mut ch = Chunk {
-      temp: temp_dir.path().to_str().unwrap().to_owned(),
-      index: 1,
-      input: Input::Video {
-        path: PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-files/blank_1080p.mkv"),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 0,
-      end_frame: 5,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::svt_av1,
-      noise_size: (None, None),
-      ignore_frame_mismatch: false,
-    };
-
-    ch.apply_photon_noise_args(None, false)?;
-    assert!(!ch.video_params.iter().any(|p| p.contains("fgs-table")));
-    Ok(())
-  }
-
-  #[test]
-  fn test_apply_photon_noise_args_unsupported_encoder() -> anyhow::Result<()> {
-    let temp_dir = tempfile::tempdir()?;
-    let mut ch = Chunk {
-      temp: temp_dir.path().to_str().unwrap().to_owned(),
-      index: 1,
-      input: Input::Video {
-        path: PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-files/blank_1080p.mkv"),
-      },
-      source_cmd: vec!["".into()],
-      output_ext: "ivf".to_owned(),
-      start_frame: 0,
-      end_frame: 5,
-      frame_rate: 30.0,
-      tq_cq: None,
-      passes: 1,
-      video_params: vec![],
-      encoder: Encoder::x264,
-      noise_size: (Some(1920), Some(1080)),
-      ignore_frame_mismatch: false,
-    };
-
-    assert!(ch.apply_photon_noise_args(Some(8), true).is_err());
     Ok(())
   }
 }
