+++ av1an-core/src/concat.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use std::fmt::{Display, Write as FmtWrite};
 use std::fs::{self, DirEntry, File};
 use std::io::Write;
@@ -360,34 +363,4 @@ pub fn ffmpeg(temp: &Path, output: &Path) -> anyhow::Result<()> {
   }
 
   Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-  use av_format::rational::Rational64;
-
-  #[test]
-  fn test_mkvmerge_options_json_no_audio() {
-    let result = mkvmerge_options_json(2, Encoder::aom, "output.mkv", None, Rational64::new(30, 1));
-    assert_eq!(
-      result,
-      r#"["-o", "output.mkv", "--default-duration", "0:30/1fps", "[", "00000.ivf", "00001.ivf","]"]"#
-    );
-  }
-
-  #[test]
-  fn test_mkvmerge_options_json_with_audio() {
-    let result = mkvmerge_options_json(
-      2,
-      Encoder::aom,
-      "output.mkv",
-      Some("audio.mkv"),
-      Rational64::new(30, 1),
-    );
-    assert_eq!(
-      result,
-      r#"["-o", "output.mkv", "audio.mkv", "--default-duration", "0:30/1fps", "[", "00000.ivf", "00001.ivf","]"]"#
-    );
-  }
 }
