+++ av1an-core/src/concat/tests.rs
@@ -1,0 +1,26 @@
+use super::*;
+use av_format::rational::Rational64;
+
+#[test]
+fn test_mkvmerge_options_json_no_audio() {
+  let result = mkvmerge_options_json(2, Encoder::aom, "output.mkv", None, Rational64::new(30, 1));
+  assert_eq!(
+    result,
+    r#"["-o", "output.mkv", "--default-duration", "0:30/1fps", "[", "00000.ivf", "00001.ivf","]"]"#
+  );
+}
+
+#[test]
+fn test_mkvmerge_options_json_with_audio() {
+  let result = mkvmerge_options_json(
+    2,
+    Encoder::aom,
+    "output.mkv",
+    Some("audio.mkv"),
+    Rational64::new(30, 1),
+  );
+  assert_eq!(
+    result,
+    r#"["-o", "output.mkv", "audio.mkv", "--default-duration", "0:30/1fps", "[", "00000.ivf", "00001.ivf","]"]"#
+  );
+}
