+++ av1an-core/src/encoder.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use std::borrow::Cow;
 use std::cmp;
 use std::fmt::Display;
@@ -49,29 +52,6 @@ pub(crate) fn parse_svt_av1_version(version: &[u8]) -> Option<(u32, u32, u32)> {
     Some((major, minor, patch))
   } else {
     None
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use crate::encoder::parse_svt_av1_version;
-
-  #[test]
-  fn svt_av1_parsing() {
-    let test_cases = [
-      ("SVT-AV1 v0.8.7-333-g010c1881 (release)", Some((0, 8, 7))),
-      ("SVT-AV1 v0.9.0-dirty (debug)", Some((0, 9, 0))),
-      ("SVT-AV1 v1.2.0 (release)", Some((1, 2, 0))),
-      ("SVT-AV1 v3.2.1 (release)", Some((3, 2, 1))),
-      ("SVT-AV1 v3.2.11 (release)", Some((3, 2, 11))),
-      ("SVT-AV1 v0.8.11 (release)", Some((0, 8, 11))),
-      ("SVT-AV1 v0.8.11-333-g010c1881 (release)", Some((0, 8, 11))),
-      ("invalid", None),
-    ];
-
-    for (s, ans) in test_cases {
-      assert_eq!(parse_svt_av1_version(s.as_bytes()), ans);
-    }
   }
 }
